organize unit test
